requirements.txt
aerich==0.7.1
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
aiosqlite==0.17.0
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
charset-normalizer==2.1.1
click==8.1.3
dictdiffer==0.9.0
frozenlist==1.3.3
idna==3.4
iso8601==1.1.0
multidict==6.0.2
numpy==1.23.4
pandas==1.5.1
pydantic==1.10.2
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0
soupsieve==2.3.2.post1
tomlkit==0.11.6
tortoise-orm==0.19.2
typing_extensions==4.4.0
yarl==1.8.1
